backend/pyproject.toml
[tool.poetry]
name = "uc-covet-backend"
version = "0.1.0"
description = "UC Covet Hardware Management System - Backend"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.6"
strawberry-graphql = {version = "0.296.1", extras = ["fastapi"]}
sqlalchemy = "2.0.36"
alembic = "1.14.1"
psycopg2-binary = "2.9.10"
python-dotenv = "1.0.1"
uvicorn = {version = "0.34.0", extras = ["standard"]}

[tool.poetry.group.dev.dependencies]
ruff = ">=0.9.0"
pytest = ">=8.0.0"
httpx = ">=0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = [
    "UP042",  # StrEnum migration risky with DB-stored enum values
    "E712",   # == False/True is correct for SQLAlchemy column expressions
]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*" = ["E501"]  # migration files have long auto-generated lines

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
